modify : 숫자형 타입 long으로 변경

+++ src/main/java/calculator/model/calculator/SumCalculator.java
@@ -3,8 +3,8 @@ package calculator.model.calculator;
 public class SumCalculator implements Calculator {
 
     @Override
-    public int calculate(int [] numbers) {
-        int answer = 0;
+    public long calculate(long [] numbers) {
+        long answer = 0;
         for(int i=0; i<numbers.length; i++) {
             answer+=numbers[i];
         }
